+++ tests/Integration/IntegrationRegistryTest.php
@@ -298,6 +298,46 @@ final class IntegrationRegistryTest extends TestCase
                 ModulesIntegration::class => new ModulesIntegration(),
             ],
         ];
+
+        yield 'Default integrations and no error handler integrations' => [
+            new Options([
+                'dsn' => 'http://public@example.com/sentry/1',
+                'default_integrations' => true,
+                'error_handler_integrations' => false,
+            ]),
+            [
+                'The "Sentry\\Integration\\RequestIntegration" integration has been installed.',
+                'The "Sentry\\Integration\\TransactionIntegration" integration has been installed.',
+                'The "Sentry\\Integration\\FrameContextifierIntegration" integration has been installed.',
+                'The "Sentry\\Integration\\EnvironmentIntegration" integration has been installed.',
+                'The "Sentry\\Integration\\ModulesIntegration" integration has been installed.',
+            ],
+            [
+                RequestIntegration::class => new RequestIntegration(),
+                TransactionIntegration::class => new TransactionIntegration(),
+                FrameContextifierIntegration::class => new FrameContextifierIntegration(),
+                EnvironmentIntegration::class => new EnvironmentIntegration(),
+                ModulesIntegration::class => new ModulesIntegration(),
+            ],
+        ];
+
+        yield 'No default integrations and error handler integrations' => [
+            new Options([
+                'dsn' => 'http://public@example.com/sentry/1',
+                'default_integrations' => false,
+                'error_handler_integrations' => true,
+            ]),
+            [
+                'The "Sentry\\Integration\\ExceptionListenerIntegration" integration has been installed.',
+                'The "Sentry\\Integration\\ErrorListenerIntegration" integration has been installed.',
+                'The "Sentry\\Integration\\FatalErrorListenerIntegration" integration has been installed.',
+            ],
+            [
+                ExceptionListenerIntegration::class => new ExceptionListenerIntegration(),
+                ErrorListenerIntegration::class => new ErrorListenerIntegration(),
+                FatalErrorListenerIntegration::class => new FatalErrorListenerIntegration(),
+            ],
+        ];
     }
 
     /**